backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
aiosqlite>=0.19.0
alembic>=1.13.0,<2.0.0

# Pydantic
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# AI clients
google-genai>=1.0.0
anthropic>=0.20.0

# Document generation
docxtpl>=0.16.0,<1.0.0
python-docx>=1.0.0,<2.0.0

# Local file storage
aiofiles>=23.0.0

# PDF processing
pypdf>=4.0.0,<5.0.0

# Authentication
pyjwt>=2.8.0,<3.0.0
httpx>=0.26.0,<1.0.0

# Stripe
stripe>=8.0.0,<10.0.0

# Utilities
python-dateutil>=2.8.0,<3.0.0
tenacity>=8.2.0,<9.0.0
